test: extract OTLP test-agent helpers from OpenTelemetrySdkTests
